End liveblog test and make the winner the default
