--- static/src/javascripts/projects/common/views/acquisitions-epic-liveblog-old-design-minimal.html
@@ -1,20 +1,0 @@
-<div class="block block--content is-liveblog-epic" data-component="<%=componentName%>">
-    <p class="block-time published-time">
-        <a href="#" itemprop="url" class="block-time__link">
-            <time data-relativeformat="med" itemprop="datePublished" class="js-timestamp"></time>
-            <span class="block-time__absolute"></span>
-        </a>
-    </p>
-    <div class="block-elements block-elements--no-byline">
-        <p>
-            <em>
-                <%=copy.heading%> <%=copy.p1%>
-            </em>
-        </p>
-        <p>
-            <em>
-                <%=copy.p2%>
-            </em>
-        </p>
-    </div>
-</div>
