--- static/src/javascripts/projects/common/views/acquisitions-epic-liveblog-old-design-subtle.html
@@ -1,19 +1,0 @@
-<div class="block block--content is-liveblog-epic is-key-event" data-component="<%=componentName%>">
-    <p class="block-time published-time">
-        <a href="#" itemprop="url" class="block-time__link">
-            <time data-relativeformat="med" itemprop="datePublished" class="js-timestamp"></time>
-            <span class="block-time__absolute"></span>
-        </a>
-    </p>
-    <div class="block-elements block-elements--no-byline">
-        <h2 class="block-title">
-            <%=copy.heading%>
-        </h2>
-        <p class="contributions__paragraph contributions__paragraph--epic">
-            <%=copy.p1%>
-        </p>
-        <p class="contributions__paragraph contributions__paragraph--epic">
-            <%=copy.p2%>
-        </p>
-    </div>
-</div>
