+++ static/src/javascripts/projects/common/views/acquisitions-epic-liveblog.html
@@ -1,32 +1,20 @@
-<div class="block block--content contributions__epic contributions__epic--liveblog" data-component="<%=componentName%>">
-    <div>
-        <h2 class="contributions__title contributions__title--epic">
-            <%=copy.heading%>
-
-        </h2>
-        <p class="contributions__paragraph contributions__paragraph--epic">
-            <%=copy.p1%>
+<div class="block block--content is-liveblog-epic" data-component="<%=componentName%>">
+    <p class="block-time published-time">
+        <a href="#" itemprop="url" class="block-time__link">
+            <time data-relativeformat="med" itemprop="datePublished" class="js-timestamp"></time>
+            <span class="block-time__absolute"></span>
+        </a>
+    </p>
+    <div class="block-elements block-elements--no-byline">
+        <p>
+            <em>
+                <%=copy.p1%>
+            </em>
         </p>
-        <p class="contributions__paragraph contributions__paragraph--epic">
-            <%=copy.p2%>
+        <p>
+            <em>
+                <%=copy.p2%>
+            </em>
         </p>
-    </div>
-
-    <div class="contributions__amount-field">
-        <div>
-            <a class="contributions__option-button contributions__contribute contributions__contribute--epic contributions__contribute--epic-member contributions__contribute--epic-member-top"
-               href="<%=membershipUrl%>"
-               target="_blank">
-                Become a supporter
-            </a>
-        </div>
-
-        <div>
-            <a class="contributions__option-button contributions__contribute contributions__contribute--epic contributions__contribute--epic-member"
-               href="<%=contributionUrl%>"
-               target="_blank">
-                Make a contribution
-            </a>
-        </div>
     </div>
 </div>
